+++ src/components/PlaylistListEdit/PlaylistListEdit.tsx
@@ -3,13 +3,13 @@ import { CardContentContainerComment } from "../common/Comments/Comments-styles"
 import {
   CrowdfundSubTitle,
   CrowdfundSubTitleRow,
-} from "../UploadVideo/Upload-styles";
+} from "../PublishFile/Upload-styles.tsx";
 import { Box, Button, Input, Typography, useTheme } from "@mui/material";
 import { useNavigate } from "react-router-dom";
 import DeleteOutlineIcon from "@mui/icons-material/DeleteOutline";
 import { removeVideo } from "../../state/features/videoSlice";
 import AddIcon from '@mui/icons-material/Add';
-import { QTUBE_VIDEO_BASE } from "../../constants";
+import { QSHARE_FILE_BASE } from "../../constants/Identifiers.ts";
 import { useSelector } from "react-redux";
 import { RootState } from "../../state/store";
 export const PlaylistListEdit = ({ playlistData, removeVideo, addVideo }) => {
@@ -21,7 +21,7 @@ export const PlaylistListEdit = ({ playlistData, removeVideo, addVideo }) => {
 const [filterSearch, setFilterSearch] = useState("")
   const search = async()=> {
 
-    const url = `/arbitrary/resources/search?mode=ALL&service=DOCUMENT&mode=ALL&identifier=${QTUBE_VIDEO_BASE}&title=${filterSearch}&limit=20&includemetadata=true&reverse=true&name=${username}&exactmatchnames=true&offset=0`
+    const url = `/arbitrary/resources/search?mode=ALL&service=DOCUMENT&mode=ALL&identifier=${QSHARE_FILE_BASE}&title=${filterSearch}&limit=20&includemetadata=true&reverse=true&name=${username}&exactmatchnames=true&offset=0`
     const response = await fetch(url, {
       method: 'GET',
       headers: {
