+++ src/components/PublishFile/PublishFile.tsx
@@ -15,7 +15,9 @@ import {
 } from "./Upload-styles";
 import {
   Box,
+  Button,
   FormControl,
+  Input,
   InputLabel,
   MenuItem,
   Modal,
@@ -26,11 +28,11 @@ import {
   useTheme,
 } from "@mui/material";
 import RemoveIcon from "@mui/icons-material/Remove";
-
 import ShortUniqueId from "short-unique-id";
 import { useDispatch, useSelector } from "react-redux";
 import AddBoxIcon from "@mui/icons-material/AddBox";
 import { useDropzone } from "react-dropzone";
+import AddIcon from "@mui/icons-material/Add";
 
 import { setNotification } from "../../state/features/notificationsSlice";
 import { objectToBase64, uint8ArrayToBase64 } from "../../utils/toBase64";
@@ -39,16 +41,26 @@ import {
   upsertVideosBeginning,
   addToHashMap,
   upsertVideos,
-  setEditVideo,
-  updateVideo,
-  updateInHashMap,
 } from "../../state/features/videoSlice";
 import ImageUploader from "../common/ImageUploader";
-import { QTUBE_VIDEO_BASE, categories, subCategories,   subCategories2,
-  subCategories3, } from "../../constants";
+import {
+  QSHARE_PLAYLIST_BASE,
+  QSHARE_FILE_BASE,
+
+
+
+
+} from "../../constants/Identifiers.ts";
 import { MultiplePublish } from "../common/MultiplePublish/MultiplePublish";
+import {
+  CrowdfundSubTitle,
+  CrowdfundSubTitleRow,
+} from "../EditPlaylist/Upload-styles.tsx";
+import { CardContentContainerComment } from "../common/Comments/Comments-styles";
 import { TextEditor } from "../common/TextEditor/TextEditor";
 import { extractTextFromHTML } from "../common/TextEditor/utils";
+import {categories, subCategories, subCategories2, subCategories3} from "../../constants/Categories.ts";
+import {titleFormatter} from "../../constants/Misc.ts";
 
 const uid = new ShortUniqueId();
 const shortuid = new ShortUniqueId({ length: 5 });
@@ -67,29 +79,30 @@ interface VideoFile {
   title: string;
   description: string;
   coverImage?: string;
-  identifier?:string;
-  filename?:string
 }
-export const EditVideo = () => {
+export const PublishFile = ({ editId, editContent }: NewCrowdfundProps) => {
   const theme = useTheme();
   const dispatch = useDispatch();
+  const [isOpenMultiplePublish, setIsOpenMultiplePublish] = useState(false);
   const username = useSelector((state: RootState) => state.auth?.user?.name);
   const userAddress = useSelector(
     (state: RootState) => state.auth?.user?.address
   );
-  const editVideoProperties = useSelector(
-    (state: RootState) => state.video.editVideoProperties
-  );
-  const [publishes, setPublishes] = useState<any[]>([]);
-  const [isOpenMultiplePublish, setIsOpenMultiplePublish] = useState(false);
-  const [videoPropertiesToSetToRedux, setVideoPropertiesToSetToRedux] =
-    useState(null);
+  const [files, setFiles] = useState<VideoFile[]>([]);
 
+  const [isOpen, setIsOpen] = useState<boolean>(false);
   const [title, setTitle] = useState<string>("");
   const [description, setDescription] = useState<string>("");
-  const [coverImage, setCoverImage] = useState<string>("");
-  const [file, setFile] = useState(null);
-  const [files, setFiles] = useState<VideoFile[]>([]);
+  const [step, setStep] = useState<string>("videos");
+  const [playlistCoverImage, setPlaylistCoverImage] = useState<null | string>(
+    null
+  );
+  const [selectExistingPlaylist, setSelectExistingPlaylist] =
+    useState<any>(null);
+  const [playlistTitle, setPlaylistTitle] = useState<string>("");
+  const [playlistDescription, setPlaylistDescription] = useState<string>("");
+  const [selectedCategory, setSelectedCategory] = useState<any>(null);
+  const [selectedSubCategory, setSelectedSubCategory] = useState<any>(null);
 
   const [selectedCategoryVideos, setSelectedCategoryVideos] =
     useState<any>(null);
@@ -99,161 +112,64 @@ export const EditVideo = () => {
     useState<any>(null);
     const [selectedSubCategoryVideos3, setSelectedSubCategoryVideos3] =
     useState<any>(null);
+    
+  const [playlistSetting, setPlaylistSetting] = useState<null | string>(null);
+  const [publishes, setPublishes] = useState<any[]>([]);
+  const { getRootProps, getInputProps } = useDropzone({
+    maxFiles: 10,
+    maxSize: 419430400, // 400 MB in bytes
+    onDrop: (acceptedFiles, rejectedFiles) => {
+      const formatArray = acceptedFiles.map((item) => {
+        return {
+          file: item,
+          title: "",
+          description: "",
+          coverImage: "",
+        };
+      });
 
-    const { getRootProps, getInputProps } = useDropzone({
-      maxFiles: 10,
-      maxSize: 419430400, // 400 MB in bytes
-      onDrop: (acceptedFiles, rejectedFiles) => {
-        const formatArray = acceptedFiles.map((item) => {
-          return {
-            file: item,
-            title: "",
-            description: "",
-            coverImage: "",
-          };
+      setFiles((prev) => [...prev, ...formatArray]);
+
+      let errorString = null;
+      rejectedFiles.forEach(({ file, errors }) => {
+        errors.forEach((error) => {
+          if (error.code === "file-too-large") {
+            errorString = "File must be under 400mb";
+          }
+          console.log(`Error with file ${file.name}: ${error.message}`);
         });
-  
-        setFiles((prev) => [...prev, ...formatArray]);
-  
-        let errorString = null;
-        rejectedFiles.forEach(({ file, errors }) => {
-          errors.forEach((error) => {
-            if (error.code === "file-too-large") {
-              errorString = "File must be under 400mb";
-            }
-            console.log(`Error with file ${file.name}: ${error.message}`);
-          });
-        });
-        if (errorString) {
-          const notificationObj = {
-            msg: errorString,
-            alertType: "error",
-          };
-  
-          dispatch(setNotification(notificationObj));
-        }
-      },
-    });
+      });
+      if (errorString) {
+        const notificationObj = {
+          msg: errorString,
+          alertType: "error",
+        };
 
-  // useEffect(() => {
-  //   if (editVideoProperties) {
-  //     const descriptionString = editVideoProperties?.description || "";
-  //     // Splitting the string at the asterisks
-  //     const parts = descriptionString.split("**");
-
-  //     // The part within the asterisks
-  //     const extractedString = parts[1];
-
-  //     // The part after the last asterisks
-  //     const description = parts[2] || ""; // Using '|| '' to handle cases where there is no text after the last **
-  //     setTitle(editVideoProperties?.title || "");
-  //     setDescription(editVideoProperties?.fullDescription || "");
-  //     setCoverImage(editVideoProperties?.videoImage || "");
-
-  //     // Split the extracted string into key-value pairs
-  //     const keyValuePairs = extractedString.split(";");
-
-  //     // Initialize variables to hold the category and subcategory values
-  //     let category, subcategory;
-
-  //     // Loop through each key-value pair
-  //     keyValuePairs.forEach((pair) => {
-  //       const [key, value] = pair.split(":");
-
-  //       // Check the key and assign the value to the appropriate variable
-  //       if (key === "category") {
-  //         category = value;
-  //       } else if (key === "subcategory") {
-  //         subcategory = value;
-  //       }
-  //     });
-
-  //     if(category){
-  //       const selectedOption = categories.find((option) => option.id === +category);
-  //   setSelectedCategoryVideos(selectedOption || null);
-  //     }
-
-  //     if(subcategory){
-  //       const selectedOption = categories.find((option) => option.id === +subcategory);
-  //   setSelectedCategoryVideos(selectedOption || null);
-  //     }
-
-  //   }
-  // }, [editVideoProperties]);
+        dispatch(setNotification(notificationObj));
+      }
+    },
+  });
 
   useEffect(() => {
-    if (editVideoProperties) {
-      setTitle(editVideoProperties?.title || "");
-      setFiles(editVideoProperties?.files || [])
-      if(editVideoProperties?.htmlDescription){
-        setDescription(editVideoProperties?.htmlDescription);
-
-      } else if(editVideoProperties?.fullDescription) {
-        const paragraph = `<p>${editVideoProperties?.fullDescription}</p>`
-        setDescription(paragraph);
-
-      }
-
-      if (editVideoProperties?.category) {
-        const selectedOption = categories.find(
-          (option) => option.id === +editVideoProperties.category
-        );
-        setSelectedCategoryVideos(selectedOption || null);
-      }
-      if (
-        editVideoProperties?.category &&
-        editVideoProperties?.subcategory &&
-        subCategories[+editVideoProperties?.category]
-      ) {
-        const selectedOption = subCategories[
-          +editVideoProperties?.category
-        ]?.find((option) => option.id === +editVideoProperties.subcategory);
-        setSelectedSubCategoryVideos(selectedOption || null);
-      }
-      if (
-        editVideoProperties?.category &&
-        editVideoProperties?.subcategory2 &&
-        subCategories2[+editVideoProperties?.subcategory]
-      ) {
-        const selectedOption = subCategories2[
-          +editVideoProperties?.subcategory
-        ]?.find((option) => option.id === +editVideoProperties.subcategory2);
-        setSelectedSubCategoryVideos2(selectedOption || null);
-      }
-      if (
-        editVideoProperties?.category &&
-        editVideoProperties?.subcategory3 &&
-        subCategories3[+editVideoProperties?.subcategory2]
-      ) {
-
-        const selectedOption = subCategories3[
-          +editVideoProperties?.subcategory2
-        ]?.find((option) => option.id === +editVideoProperties.subcategory3);
-        setSelectedSubCategoryVideos3(selectedOption || null);
-      }
-
-      
+    if (editContent) {
     }
-  }, [editVideoProperties]);
+  }, [editContent]);
 
   const onClose = () => {
-    dispatch(setEditVideo(null));
-    setVideoPropertiesToSetToRedux(null);
-    setFile(null);
-    setTitle("");
-    setDescription("");
-    setCoverImage("");
+    setIsOpen(false);
   };
 
-  async function publishQDNResource() {
-    try {
-      if (!title) throw new Error("Please enter a title");
-      if (!description) throw new Error("Please enter a description");
-      if (!selectedCategoryVideos) throw new Error("Please select a category");
-      if (!editVideoProperties) return;
-      if (!userAddress) throw new Error("Unable to locate user address");
-      if(files.length === 0) throw new Error("Add at least one file");
 
+
+  async function publishQDNResource() {
+
+      try {
+        if (!userAddress) throw new Error("Unable to locate user address");
+
+        if (!title) throw new Error("Please enter a title");
+        if (!description) throw new Error("Please enter a description");
+        if (!selectedCategoryVideos) throw new Error("Please select a category");
+        if(files.length === 0) throw new Error("Add at least one file");
       let errorMsg = "";
       let name = "";
       if (username) {
@@ -264,7 +180,7 @@ export const EditVideo = () => {
           "Cannot publish without access to your name. Please authenticate.";
       }
 
-      if (editVideoProperties?.user !== username) {
+      if (editId && editContent?.user !== name) {
         errorMsg = "Cannot publish another user's resource";
       }
 
@@ -277,32 +193,30 @@ export const EditVideo = () => {
         );
         return;
       }
+
       let fileReferences = []
 
       let listOfPublishes = [];
-      const fullDescription = extractTextFromHTML(description);
 
-      const category = selectedCategoryVideos.id;
-      const subcategory = selectedSubCategoryVideos?.id || "";
-      const subcategory2 = selectedSubCategoryVideos2?.id || "";
-      const subcategory3 = selectedSubCategoryVideos3?.id || "";
-      const sanitizeTitle = title
-      .replace(/[^a-zA-Z0-9\s-]/g, "")
-      .replace(/\s+/g, "-")
-      .replace(/-+/g, "-")
-      .trim()
-      .toLowerCase();
+        const fullDescription = extractTextFromHTML(description);
+        const category = selectedCategoryVideos.id;
+        const subcategory = selectedSubCategoryVideos?.id || "";
+        const subcategory2 = selectedSubCategoryVideos2?.id || "";
+        const subcategory3 = selectedSubCategoryVideos3?.id || "";
 
+        const sanitizeTitle = title
+          .replace(/[^a-zA-Z0-9\s-]/g, "")
+          .replace(/\s+/g, "-")
+          .replace(/-+/g, "-")
+          .trim()
+          .toLowerCase();
 
       for (const publish of files) {
-        if(publish?.identifier){
-          fileReferences.push(publish)
-          continue
-        }
+     
         const file = publish.file;
         const id = uid();
 
-        const identifier = `${QTUBE_VIDEO_BASE}${sanitizeTitle.slice(0, 30)}_${id}`;
+        const identifier = `${QSHARE_FILE_BASE}${sanitizeTitle.slice(0, 30)}_${id}`;
 
         
 
@@ -337,6 +251,7 @@ export const EditVideo = () => {
         let metadescription =
         `**cat:${category};sub:${subcategory};sub2:${subcategory2};sub3:${subcategory3}**` +
           fullDescription.slice(0, 150);
+
      
         const requestBodyVideo: any = {
           action: "PUBLISH_QDN_RESOURCE",
@@ -347,7 +262,7 @@ export const EditVideo = () => {
           description: metadescription,
           identifier,
           filename,
-          tag1: QTUBE_VIDEO_BASE,
+          tag1: QSHARE_FILE_BASE,
         };
         listOfPublishes.push(requestBodyVideo);
         fileReferences.push({
@@ -360,12 +275,14 @@ export const EditVideo = () => {
         })
       }
 
+      const idMeta = uid();
+      const identifier = `${QSHARE_FILE_BASE}${sanitizeTitle.slice(0, 30)}_${idMeta}`;
       const fileObject: any = {
         title,
-        version: editVideoProperties.version,
+        version: 1,
         fullDescription,
         htmlDescription: description,
-        commentsId: editVideoProperties.commentsId,
+        commentsId: `${QSHARE_FILE_BASE}_cm_${idMeta}`,
         category,
         subcategory,
         subcategory2,
@@ -379,7 +296,6 @@ export const EditVideo = () => {
 
       const crowdfundObjectToBase64 = await objectToBase64(fileObject);
       // Description is obtained from raw data
-
       const requestBodyJson: any = {
         action: "PUBLISH_QDN_RESOURCE",
         name: name,
@@ -387,58 +303,37 @@ export const EditVideo = () => {
         data64: crowdfundObjectToBase64,
         title: title.slice(0, 50),
         description: metadescription,
-        identifier: editVideoProperties.id,
-        tag1: QTUBE_VIDEO_BASE,
+        identifier: identifier + "_metadata",
+        tag1: QSHARE_FILE_BASE,
         filename: `video_metadata.json`,
       };
       listOfPublishes.push(requestBodyJson);
-
       setPublishes(listOfPublishes);
       setIsOpenMultiplePublish(true);
-      setVideoPropertiesToSetToRedux({
-        ...editVideoProperties,
-        ...fileObject,
-      });
     } catch (error: any) {
       let notificationObj: any = null;
       if (typeof error === "string") {
         notificationObj = {
-          msg: error || "Failed to publish update",
+          msg: error || "Failed to publish share",
           alertType: "error",
         };
       } else if (typeof error?.error === "string") {
         notificationObj = {
-          msg: error?.error || "Failed to publish update",
+          msg: error?.error || "Failed to publish share",
           alertType: "error",
         };
       } else {
         notificationObj = {
-          msg: error?.message || "Failed to publish update",
+          msg: error?.message || "Failed to publish share",
           alertType: "error",
         };
       }
       if (!notificationObj) return;
       dispatch(setNotification(notificationObj));
-
-      throw new Error("Failed to publish update");
     }
   }
 
-  const handleOnchange = (index: number, type: string, value: string) => {
-    // setFiles((prev) => {
-    //   let formattedValue = value
-    //   console.log({type})
-    //   if(type === 'title'){
-    //     formattedValue = value.replace(/[^a-zA-Z0-9\s]/g, "")
-    //   }
-    //   const copyFiles = [...prev];
-    //   copyFiles[index] = {
-    //     ...copyFiles[index],
-    //     [type]: formattedValue,
-    //   };
-    //   return copyFiles;
-    // });
-  };
+
 
   const handleOptionCategoryChangeVideos = (
     event: SelectChangeEvent<string>
@@ -457,7 +352,6 @@ export const EditVideo = () => {
     );
     setSelectedSubCategoryVideos(selectedOption || null);
   };
-
 
   const handleOptionSubCategoryChangeVideos2 = (
     event: SelectChangeEvent<string>,
@@ -482,10 +376,27 @@ export const EditVideo = () => {
   };
 
 
+
   return (
     <>
+      {username && (
+        <>
+          {editId ? null : (
+            <StyledButton
+              color="primary"
+              startIcon={<AddBoxIcon />}
+              onClick={() => {
+                setIsOpen(true);
+              }}
+            >
+              share
+            </StyledButton>
+          )}
+        </>
+      )}
+
       <Modal
-        open={!!editVideoProperties}
+        open={isOpen}
         aria-labelledby="modal-title"
         aria-describedby="modal-description"
       >
@@ -497,24 +408,29 @@ export const EditVideo = () => {
               justifyContent: "space-between",
             }}
           >
-            <NewCrowdfundTitle>Update share</NewCrowdfundTitle>
+          
+              <NewCrowdfundTitle>Share</NewCrowdfundTitle>
+       
           </Box>
-          <>
-            <Box
-              {...getRootProps()}
-              sx={{
-                border: "1px dashed gray",
-                padding: 2,
-                textAlign: "center",
-                marginBottom: 2,
-                cursor: "pointer",
-              }}
-            >
-              <input {...getInputProps()} />
-              <Typography>Click to add more files</Typography>
-            </Box>
-            {files.map((file, index) => {
-              const isExistingFile = !!file?.identifier
+
+          {step === "videos" && (
+            <>
+              <Box
+                {...getRootProps()}
+                sx={{
+                  border: "1px dashed gray",
+                  padding: 2,
+                  textAlign: "center",
+                  marginBottom: 2,
+                  cursor: "pointer",
+                }}
+              >
+                <input {...getInputProps()} />
+                <Typography>
+                  Drag and drop files here or click to select files
+                </Typography>
+              </Box>
+              {files.map((file, index) => {
                 return (
                   <React.Fragment key={index}>
                     <Box
@@ -524,7 +440,7 @@ export const EditVideo = () => {
                         alignItems: "center",
                       }}
                     >
-                      <Typography>{isExistingFile? file.filename : file?.file?.name}</Typography>
+                      <Typography>{file?.file?.name}</Typography>
                       <RemoveIcon
                         onClick={() => {
                           setFiles((prev) => {
@@ -541,15 +457,14 @@ export const EditVideo = () => {
                   </React.Fragment>
                 );
               })}
-           
-            <Box
-              sx={{
-                display: "flex",
-                gap: "20px",
-                alignItems: "flex-start",
-              }}
-            >
-               {files?.length > 0 && (
+              <Box
+                sx={{
+                  display: "flex",
+                  gap: "20px",
+                  alignItems: "flex-start",
+                }}
+              >
+                {files?.length > 0 && (
                   <>
                    <Box sx={{
                       display: 'flex',
@@ -676,8 +591,8 @@ export const EditVideo = () => {
                    
                   </>
                 )}
-            </Box>
-            {files?.length > 0 && (
+              </Box>
+              {files?.length > 0 && (
                 <>
                   <CustomInputField
                     name="title"
@@ -686,10 +601,7 @@ export const EditVideo = () => {
                     value={title}
                     onChange={(e) => {
                       const value = e.target.value;
-                      const formattedValue = value.replace(
-                        /[^a-zA-Z0-9\s-_!?]/g,
-                        ""
-                      );
+                      const formattedValue = value.replace(titleFormatter, "");
                       setTitle(formattedValue);
                     }}
                     inputProps={{ maxLength: 180 }}
@@ -710,8 +622,9 @@ export const EditVideo = () => {
                   />
                 </>
               )}
-          </>
-
+            
+            </>
+          )}
           <CrowdfundActionButtonRow>
             <CrowdfundActionButton
               onClick={() => {
@@ -741,21 +654,29 @@ export const EditVideo = () => {
           </CrowdfundActionButtonRow>
         </ModalBody>
       </Modal>
+
       {isOpenMultiplePublish && (
         <MultiplePublish
           isOpen={isOpenMultiplePublish}
           onSubmit={() => {
             setIsOpenMultiplePublish(false);
-            const clonedCopy = structuredClone(videoPropertiesToSetToRedux);
-            dispatch(updateVideo(clonedCopy));
-            dispatch(updateInHashMap(clonedCopy));
+            setIsOpen(false);
+            setFiles([]);
+            setStep("videos");
+            setPlaylistCoverImage(null);
+            setPlaylistTitle("");
+            setPlaylistDescription("");
+            setSelectedCategory(null);
+            setSelectedSubCategory(null);
+            setSelectedCategoryVideos(null);
+            setSelectedSubCategoryVideos(null);
+            setPlaylistSetting(null);
             dispatch(
               setNotification({
-                msg: "File updated",
+                msg: "Files published",
                 alertType: "success",
               })
             );
-            onClose();
           }}
           publishes={publishes}
         />
