+++ src/constants/Identifiers.ts
@@ -1,0 +1,17 @@
+const useTestIdentifiers = false;
+
+export const QSHARE_FILE_BASE = useTestIdentifiers
+  ? "MYTEST_share_vid_"
+  : "qshare_file_";
+
+  export const QSHARE_PLAYLIST_BASE = useTestIdentifiers
+  ? "MYTEST_share_playlist_"
+  : "qshare_playlist_";
+
+  export const QSHARE_COMMENT_BASE = useTestIdentifiers
+  ? "qcomment_v1_MYTEST_"
+  : "qcomment_v1_qshare_";
+
+
+
+
