+++ src/constants/Misc.ts
@@ -1,0 +1,1 @@
+export const titleFormatter = /[^a-zA-Z0-9\s-_!?()&'",.~;:|]/g;
