+++ src/pages/Home/Home.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { VideoList } from './VideoList'
+import { FileList } from './FileList.tsx'
 
 import { useSelector } from 'react-redux'
 import { RootState } from '../../state/store'
@@ -8,7 +8,7 @@ export const Home = () => {
 
   return (
     <>
-    <VideoList />
+    <FileList />
     </>
    
   )
